% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchNorm.R
\name{GetiLISI}
\alias{GetiLISI}
\title{Get iLISI}
\usage{
GetiLISI(object = NULL, nSamples = NULL)
}
\arguments{
\item{object}{A Seurat object with a UMAP computed, and sample identifiers stored as "orig.ident" in the object metadata slot.}

\item{nSamples}{The number of unique samples within the Seurat object (the number of samples which have been integrated in the UMAP).}
}
\value{
Returns a numeric score from 0-1 (best-worst).
}
\description{
Compute LISI scores (for each cell) for a Seurat object with a specified number of samples.
The median LISI score across all cells is then corrected for the total number of samples, and converted into a format where 0 = maximum integration and 1 = maximum segregation.
A single score on a scale of 0-1 is returned.
}
\examples{
GetiLISI(object = my_seurat_object, nSamples = 3)
}
