% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchNorm.R
\name{import_PBMCs}
\alias{import_PBMCs}
\title{Download and import single-cell data}
\usage{
import_PBMCs(sample = NULL)
}
\arguments{
\item{url}{Provide the complete file path to the}
}
\value{
Returns a pre-processed Seurat object
}
\description{
Download the dataset from GEO, filter, and create a Seurat object
}
\references{
Benjamin R. Babcock, et al.
Data Matrix Normalization and Merging Strategies Minimize Batch-specific Systemic Variation in scRNA-Seq Data
Freely available as a preprint on bioRxiv. doi: https://doi.org/10.1101/2021.08.18.456898
}
