% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchNorm.R
\name{MitoFilter}
\alias{MitoFilter}
\title{Mitochondrial Transcripts Filter}
\usage{
MitoFilter(obj = NULL)
}
\arguments{
\item{obj}{The vector of variable features used to construct the PCs}

\item{mtx.use}{The Seurat object to subset.
This object must already contain a metadata feature "percent.mito" containing the total mitochondrial percentage, by cell.
This metadata feature can be added by the "PercentageFeatureSet" function.}
}
\value{
Returns a filtered Seurat object with only cells passing the mitochondrial percentage filter, as well as containing greater than 100 unique RNA transcripts.
}
\description{
Automatically identify an appropriate threshold for subsetting a Seurat object by mitochondrial percentage (fraction of total transcripts beginning with "MT-".
MitoFilter first identifies the median mitochondrial percentage value five standard deviations above the median for a given sample.
MitoFilter then subsets a Seurat object, retaining only cells which have a lesser mitochondrial percentage than the cutoff (median +5 SD).
The object is returned containing only cells below threshold, as well as only cells with greter than 100 unique RNA transcripts.
}
\examples{
my_seurat_object <- MitoFilter(obj = my_seurat_object)
my_seurat_object <- my_seurat_object \%>\%
                    MitoFilter()
}
