% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchNorm.R
\name{splitCD}
\alias{splitCD}
\title{Split "CD" Feature Names}
\usage{
splitCD(x = NULL)
}
\arguments{
\item{x}{A vector of feature names to be corrected}
}
\value{
Returns a character vector of feature names in the desired format.
}
\description{
Split feature names to extract "CDx" from ADT labels.
splitCD converts names in the format "CDx_xxxxxx" to "CDx", while eliminating whitespace and correcting capitalization.
A vector of feature names in the corrected format is returned.
}
\examples{
splitCD("CD19_TotalSeqC")
corrected.names <- splitCD(c("CD19_TotalSeqC", "CD45_TotalSeqC"))
}
