% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchNorm.R
\name{TestPCA}
\alias{TestPCA}
\title{Test PCA}
\usage{
TestPCA(
  genes.use = object@assays$RNA@var.features,
  mtx.use = object@assays$RNA@scale.data
)
}
\arguments{
\item{genes.use}{The vector of variable features used to construct the PCs}

\item{mtx.use}{The expression matrix used to construct the PCs.
At a minimum, this matrix must included the variable features included in "genes.use"}
}
\value{
Returns a table containing the z-score of the cumulative percent of total variance for each PC
}
\description{
Identify important PCs from total principal components analysis (within a Seurat object).
TestPCA functions by generating a z-score corresponding to each respective PC's proportional contribution to the total variance.
Can be used similarly to the Seurat function ElbowPlot, which plots each successive PC by its standard deviation.
}
